agent: Avoid task confusion
Commit cb2eb0afbe49278b53a23648ae5e4cf91b3ac664 changed the agent to use
a []api.Task instead of []*api.Task. This looked like an innocent
change, but it causes very strange behavior. I think the problem is that
some functions store the task pointer we pass it. Any reads from this
pointer after the loop has advanced would read from the wrong task,
since the pointer points to the task iterator.

Changing this back to []*api.Task.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>