allocator: Batch service and node allocations
On startup, the allocator can do many separate raft writes to allocate
services and nodes. Solve this by batching those writes.

Split allocation functions into one function that does the allocation,
and another function that commits it to the store. Always do the
allocations first and then commit them in a single batch.

This also converts task allocation not to hold the raft store lock
while doing allocations, which will be important as plugin support is
added.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>