all: Avoid unnecessary "append" loops
In cases where append(a, b...) can be used, don't replicate the same
functionality with a for loop.

Modify the deepcopy plugin to generate code that uses the former syntax.

Fix an instance in allocator. Also change a slice == nil check to
len(slice) == 0, which caused some fun to debug problems.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>