api: add support for setting a container's hostname
Adds support for setting a container's hostname via the
`ContainerSpec.Hostname`. All containers in the service will get the
same value for the hostname.

This was implemented in a series of 5 commits by

Separate integration tests for CI

Signed-off-by: Nikolay Milovanov <nmil@itransformers.net>

Added option hostname to docker service create

Signed-off-by: Vasil Yordanov <vasil.yordanov@gmail.com>

Use Hostname field in agent/exec/container

Signed-off-by: Vasil Yordanov <vasil.yordanov@gmail.com>

Added validation for hostname parameter

Signed-off-by: Vasil Yordanov <vasil.yordanov@gmail.com>

Pre-compile the regexp for the hostname validation

Signed-off-by: Vasil Yordanov <vasil.yordanov@gmail.com>
Signed-off-by: Stephen J Day <stephen.day@docker.com>