api: additions for host-level port publishing
To allow exposing ports on the host to which a task is scheduled, we
have introduced the concept of an publish mode parameter to the
PortConfig. The default, ingress, works as it does today. The new mode,
host, will bind the port to the host's port space, optimistically
deferring allocation to the target node.

Signed-off-by: Stephen J Day <stephen.day@docker.com>