integration: Inject logger that differentiates nodes
It can be difficult to tell which node a log message comes from. This
adds a logger to the context that has a log field identifying which node
produced the log message.

Right now this is only somewhat useful because a lot of log messages
come from grpc and etc/raft, but in the future as more is converted to
use the context for logging it will be increasingly useful.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>