Add blacklist for removed nodes
We don't want to consider certificates valid if they were associated
with nodes that are now removed from the swarm. Add a blacklist
mechanism that keeps track of removed node IDs in the cluster object.
This also stores the expiration time of the most recently issued
certificate, though there is not yet any expiry mechanism to take
advantage of this.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>