identity: Extend random ID domain to fill all 25 characters
The current identity package generates 16 bytes of entropy to fill in 25
base-36 characters. There is a slight mismatch between these - 25
base-36 characters actually corresponds to about 129.25 bits. This means
there will always be a leading 0 bit in the generated ID. The
user-facing consequence is that IDs always start with [0-9a-f], which
makes partial IDs less effective since the first digit is much more
likely to collide.

To fix this, generate an extra byte of entropy, and render the result
into a base-36 string with one extra character. Discard the leading
character, which will have a limited range. The returned 25 characters
should have uniform values.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>