Add support for rolling restart
We don't currently have a way to trigger a restart of all tasks from a
service. A workaround is to make a change to the service using "docker
service update", but if no changes are made, no restarts will happen.
This commit adds a way to trigger a (possibly rolling) restart without
necessarily making any changes to the service. This should address a lot
of user questions about how to restart services, for example to reload
configuration.

Since updates are built around a reconcilation loop, this works by
introducing a field in TaskSpec that gets changed when a restart is
forced. Thus, no changes at all are needed in the orchestrator. The
restart option still makes a change that's picked up by the
orchestrator, but it's a change that does not affect the service in any
way other than triggering a restart.

I envision this being added as an option to "docker service update". For
example:

    docker service update --force --update-parallelism 2 \
    --update-delay 5s myservice

This would allow us to reuse existing "docker service update" options
such as "--update-parallelism".

Alternatively, it could be exposed through a new "docker service
restart" subcommand:

   docker service restart --parallelism 2 --delay 5s myservice

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>