api: Merge Pending and Allocated states
The intent of the Pending state was for tasks which had finished
allocation but had not yet been scheduled. We aren't actually using this
state - we use Allocated for this purpose instead.

Rename Allocated to Pending, but keep the same enum value. This means
there is no functional change or API change, but the name of the state
is more descriptive.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>