orchestrator: Make global tests more robust; use generic Expect function
The global orchestrator tests were pausing 200 ms to wait for some
events to pass. It's more reliable to just consume these specific
events. This changes them to do that, and also replaces
testutils.Expect* with a single Expect function. This is nicer than
having a separate function for every event type. The Checks in the event
structure can be used to match specific event properties.

In the future, I may also convert testutils.Watch* to a similar Watch
function.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>