Provide a mTLS CA server endpoint to get the latest manager unlock key so that so that TLS certificates are never written to disk after they are downloaded from the swarm CA.
Also, update the cluster spec and object so that a user can enable and disable autolocking
managers. If autolocking is enabled, a shared key will be used by all the managers to
encrypt TLS keys and raft DEKs.

Also, provide the control API and swarmd/swarmct commands to enable/disable autolocking,
and rotate lock keys.

Signed-off-by: cyli <ying.li@docker.com>