Basic/skeleton Process type task and ProcessSpec.
I would like to implement a new Executor which will run a simple process. That
executor does not need to live in this reprository, but the basic datatype
does. Add a very simple datatype to encapsulate the properties of a Task of
this kind. Initially its only properties are the argv to use (nb: this is
intended to be argv[1:N] rather than argv[0:N], so the Executor still retains
oversight on what binaries are run.

Signed-off-by: Ian Campbell <ian.campbell@docker.com>