secrets: decouple secrets from agent package
Previous changes tightly coupled the secrets storage with the agent
package. The evidence of this lies in the addition of an argument to
`Executor.Controller`, which typically means something isn't quite
right. Controllers should be resolvable with only the task information.

We now defer the entireity of secret storage to the executor. The
existing map-based implementation is now in a package `secrets` which
can used by executors.

No interface changes are required by upstreams. Secrets support is
declared by executor implementation via the `SecretsProvider` interface.

Signed-off-by: Stephen J Day <stephen.day@docker.com>