manager: Fix advertise address regression
The advertise address was being forced to the same value as the listen
address, by some code added to (*Manager).New for integration testing.
This code was overwriting tcpAddr (which is supposed to contain the
advertise address) with the bound address of the listening socket. The
idea was to support passing a :0 address into the manager, through node
(so a listen can't be bound beforehand). Fix this to only override
tcpAddr if a 0 port was actually given.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>