manager: Clean up address configuration
Previously, we had a ProtoAddr map and a ProtoListener map. These maps
were awkward because keying on the protocol doesn't really make sense.

Replace these with RemoteAPI and ControlAPI fields. RemoteAPI is a
struct that contains a listening address and an optional advertise
address. ControlAPI is just an address string, because the concept of
advertise address does not apply to the control API.

Remove the capability to pass in listeners. This was only used by a
single unit test. The concept can't extend to node, because node may
start the manager multiple times, and a listener can't be reused.
Change the unit test to pass in an address with a 0 port and let the
manager bind.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>