Make port allocator not depend on order of ports
Currently the port allocator depends on the order of the ports in the
Spec and the Endpoint to be the same to determine if it is allocated or
while reconciling ports. This results in unnecessary false negatives
while testing for allocated and go through unnecessary
allocations. Fixed this by treating the list of ports as a set and check
if the same set of ports are available in the Endpoint as in Spec.

Signed-off-by: Jana Radhakrishnan <mrjana@docker.com>