api: add support for security profiles
We define the basis for a profile-based security dispatch. We put
forward a model, based on existing `--privileged` flag that can be
extended in the future to support more complex workflows.

The first aspect is to allow the container runtime to select a security
profile, keyed by name. We introduce two profiles, "default" and
"privileged" corresponding to the currently available control we have
today. We will extend this model to support more complex scenarios, with
more granular control, in the future.

We balance this model with a Node security profile ACL that can select
which profiles can be run on a node. Users can add and remove items from
this ACL to control which security levels can run on a given node. We
define the empty list to be allowing the "Default" security profile.

To support future interaction, we also define that profiles are
additive. This means that the capability of a set of profiles is a union
of all the specified profiles.

Signed-off-by: Stephen J Day <stephen.day@docker.com>