Add indices over tasks and services by referenced network and secret IDs
Previously, it was not possible to have indices which could have more
than one value per object. This vendors a fork of go-memdb that adds
this feature, from a PR that has been pending for more than eight
months. It then adds indices over tasks and services for the network and
secret IDs referenced by the tasks and services. Network deletion is
updated to use the new index instead of listing all services.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>