Add per-Role additional authentications to RootCA
These are added as subjectAltName DNS names which together with the previous
patch allows the application to grant specific additional authorizations to
each role.

The initial set of authorizations is specified by the application via
node.Config and is then stored in the raft data store. At this time there is no
provision for changing them on the fly (e.g. a swarmctl cluster update option).

Signed-off-by: Ian Campbell <ian.campbell@docker.com>