Update the comment for `Hosts` field in `ContainerSpec`
This fix is a followup for `Hosts` field in `ContainerSpec`,
from the comment in:
https://github.com/docker/swarmkit/pull/1729#pullrequestreview-8044827
and from docker PR:
https://github.com/docker/docker/pull/28031

In the docker PR, it has been specified that `Hosts` in `ContainerSpec`
should only follow the standard format:
```
IP_address canonical_hostname [aliases...]
```
The docker PR 28031 already takes the above into consideration so
no implementation updates needed.

This fix corrects the comment in the specification of `Hosts`.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>