api: Centralize defaults and add InsertDefaults flag
Currently, there is a problem with visibility of defaults. The best way
to create a service is to only specify fields that you care about,
because then the manager can apply appropriate defaults for the other
fields. However, there's no way to know which values are actually being
used for those fields. The spec is under the user's control, so if they
don't specify a value (for example, `Spec.Task.Restart.Delay`), this
field will show up as blank when the service is retrieved.

We don't want to break the spec ownership model, but we do want a way to
communicate actual values in use back to the client. For this, we add an
`InsertDefaults` flag to the ControlAPI which allows fetching a version
of the service with defaults inserted into empty fields.

This also has the benefit of centralizing defaults. Default values for
these fields that were defined in various places around the code are now
centralized in `github.com/docker/swarmkit/api/defaults`.

Right now this flag is only is only available for `Service`, but as we
encounter situations where other objects have default values which need
to be exposed, we should add this to other objects as well.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>