ca: Honor context in (*ExternalCA).Sign
If the context is cancelled while trying to communicate with an external
CA, the Sign method should return immediately. Without this, it can take
a long time to shut down the manager, because it waits for this method
to finish. If the external CA is not reachable, it can take especially
long.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>