Add generic top-level Resource type
SwarmKit provides a consistent distributed data store, but only native
types can be stored inside it. Frequently, other components want to
store data in this store, but there is usually no appropriate place to
put it.

This commit adds a generic top-level Resource object that can carry an
arbitrary payload. These objects have a type described by a string, to
prevent two objects carrying different kinds of payloads from colliding.
Indexing is defined by the creator of the Resource through key/value
pairs, instead of automatically deriving indices from the payload of the
object (which is opaque to the data store in this case). A subsequent PR
will add automatic indexing over protobuf content.

Indices has been added to all existing top-level types, so any type can
make use of custom indexing.

The payload of the object is a google.protobuf.Any message, which is a
standardized way of conveying an arbitrary serialized message.

Before a Resource object can be used, a corresponding type must be
registered using an Extension object.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>