Allow swarm join with `--availability=drain`
This fix (together with docker/docker#24993) tries to address the issue raised in

docker/docker#24596

where it was not possible to join as manager only (`--availability=drain`).

This fix adds Availability to IssueNodeCertificateRequest so that it is
possible to specify the availability of the node when it joins the swarm.

This fix is related to
docker/docker#24596
docker/docker#24993

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>