deepcopy: refactor output of T.Copy() T*
To make the deepcopy plugin interact better with externally defined
types, it has been refactored to define a more flexible interface, based
around `CopyFrom`. `CopyFrom` takes a `src` argument and copies `src`
the fields of the method receiver. On top of this, we can build
`deepcopy.Copy` which takes a pre-allocated `dst` and `src` argument.
Because we provide our own copy implementation, as a parameter of the
plugin, we can inject our own implementations in the copy function
without affecting valid compilation of generated code.

In particular, this will allow us to use externally defined timestamp
and duration types, in addition to other well known types, that were
previously unusable due to the requirement that the Copy() function be
generated on types.

To preserve existing direct usage of `T.Copy() *T`, we still generated
the method for use in code. Since this code is typically for direct
usage, and has access to the Go compiler when under use (as opposed to
the protoc generation stack), this usage is safe and encouraged.

Signed-off-by: Stephen J Day <stephen.day@docker.com>