Validate secret for white spaces only data
This fix is related to docker issue:
https://github.com/docker/docker/issues/28927

Currently, secret is validated and the length of
the data has to be 0 < len < MaxSecretSize.

However, sometimes data might only contain white spaces
(e.g., `" "`, `"\n"`, etc.) and the secret could be
created as well.

This fix adds the additional validation so that white spaces
only data will not be accepted.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>