Fix missing IPAM options in swarm network mode
This fix tries to fix the issue raised in docker/29044 where
the IPAM options is missing in swarm network mode after the service is deployed.
Before the service is deployed, the IPAM options is available though.

The reason is that when network allocator updates the network, the
IPAM.Options is not updated.

This fix fixes this issue.

This fix is also related to docker PR:
docker/29074.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>