logbroker: Fix race on subscription.message.Close
Encountered a data race while running the unit tests on the 1.13.x
branch. The Close field is set without locking. Make sure it is only set
and checked with the subscription's lock held.

I also considered creating a copy of the subscription before setting
Close to true and sending it over the queue, but this seems a bit more
straightforward.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>