raft: Add grace period to sendToMember
This function was recently changed to synchronize sends to the same
member, so that sending goroutines wouldn't impact each other by closing
the connection on an error.

To prevent the number of waiting goroutines from growing without bound,
SendTimeout was applied to both waiting for the last sender to finish,
and doing the actual I/O.

This may have unwanted consequences if two sends to the same member are
started around the same time and the first times out. The second will
only have a tiny amount of time to complete after waiting. It will
probably time out too, and bounce the connection as a result. So the
connection may be closed and reopened more than necessary, and this may
impact health checks, which reuse the same connections.

Change this function to reset the timeout after it finishes waiting.

I think this may help with long election loops seen in Docker Engine's
TestSwarmLockUnlockCluster. The logs from these runs showed "member
which sent vote request failed health check", which is unexpected, and
probably means a connection was being wrongly closed from underneath
ProcessRaftMessage. In case this change doesn't solve the problem, I've
added more detailed error logging that will show how the health check
failed.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>