When completing a raft DEK key rotation (which only happens when a snapshot is completed using the new DEK key, and the snapshot's index must be >= the index of the last WAL written with the old key), remove the previous snapshots and WALs, since the old DEK will no longer be available and they can't be decrypted anymore anyway.
Also, when a manager is removed from the cluster, the raft DEK will be removed and the TLS key
rotated to be that of a worker.  Since the DEK won't be available anyway, just remove the old
wal and snapshot directories instead of backing them up.

Signed-off-by: cyli <ying.li@docker.com>
(cherry picked from commit 57343aed13c65c3ba175c12536ba44dc8792846e)