Remove suffix of `.<TaskID>` from the name of `Task`
This fix is related to docker PR:
https://github.com/docker/docker/pull/29196

The above PR will remove the suffix of `.<TaskID>`
from the name of `Task`.

This fix tries to remove suffix of `.<TaskID>` from
the related SwarmKit part.

Technically, the docker PR https://github.com/docker/docker/pull/29196
is only about the "display name" of the task.
We could very well leave the "internal name" of the task alone (keep `.<TaskID>`)
if we prefer.

So please fee free to close this PR if keep `.<TaskID>` is preferred.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>