Fix duplicated ports allocation with restarted swarm
This fix fixes the issue raised in:
https://github.com/docker/docker/issues/29247

where duplicated ports (`PublishedPort`) is being allocated
with restarted swarm nodes.

The reason for the issue was that `isPortsAllocated` relies
on portSpace which does not capture the restored services when
swarm is restarted.
This fix add a flag in `isPortsAllocated` so that in case
the allocator is still in initialization stage, dynamically
allocated ports (`PublishedPort`) will always be reallocated.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>