raftproxy: Remove streaming proxy tests
These tests are inherently racy because they send to a streaming RPC
handler that will exit immediately. As a result, they were flaky. See
https://github.com/docker/swarmkit/issues/1662#issuecomment-266928666
for more details.

The tests are only covering a pathological case that can't really be
tested in this way because of the race. Removing them seems like the
best course of action.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>