raft: Prevent the same manager from joining raft twice
If a manager's raft state is lost or corrupted, it may try to join the
Raft cluster as a fresh member. This must be avoided, because it will
influence the quorum of the Raft cluster, which can prevent leader
elections from succeeding.

Add a check to Join that prevents a node with the same identity from
joining twice.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>
(cherry picked from commit a2037c38c54a028fe11687fb2b6e5d73b78aad58)