raft: Ignore messages intended for a different member
If a node managed to join raft twice, it would receive messages meant
for both the new and old raft members. There was no check verifying that
the recipient of message is correct. Adding this check makes sure that
if this situation were to arise (perhaps because the leader is an older
version that allows Join to be called twice by the same node), extra
messages meant for a different raft member wouldn't be passed through to
Step and possibly cause a panic.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>