raft: Stop manager on write error
If we failed to save some entries to durable storage, we must stop the
manager. The existing code just logs an error and continues, but this
means the state essentially becomes corrupted and the manager may panic
later. It's better to fail immediately.

The code path that involves Run returning early is not especially
robust, and we've noticed some issues on demotion related to this.
Despite this, the alternative of letting Run continue in this situation
is much worse. We will need to fix the code path where Run returns early
anyway, because it is used in demotion.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>