logbroker: Fix flaky test TestLogBrokerNoFollow
The issue appears to be that stream RPC handlers are invoked
asynchronously. When ListenSubscriptions is called and returns a stream,
we don't have a guarantee that any code inside that RPC handler has
executed yet. When we call SubscribeLogs, ListenSubscriptions may not
have called registerSubscription yet, in which case SubscribeLogs will
think the associated node is down, and will call Done for that node
automatically.

Since the current API doesn't give us a way to ensure that the node has
registered as a log listener, insert a sleep to give
ListenSusbscriptions time to do its work. Ideally there would be a
better way to do this.

Also, remove some synchronization in listenSubscriptions that doesn't
help, and remove a capacity argument to "make" in a case where only a
subset of the original slice's values are being copied.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>