manager: Allow API listeners to be bound after starting the manager
This adds BindRemote and BindControl methods to Manager, which can be
used to specify the remote API and control API addresses after creating
or starting the manager. Raft has been updated to accept a new remote
address after being started. If The RemoteAPI and ControlAPI fields are
passed to manager.New, it is not necessary to call BindRemote or
BindControl explicitly.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>