Reconcile roles with raft member list
When a node is demoted, two things need to happen: the node object
itself needs to be updated, and the raft member list needs to be updated
to remove that node. Previously, it was possible to get into a bad state
where the node had been updated but not removed from the member list.

This changes the approach so that controlapi only updates the node
object, and there is a goroutine that watches for node changes and
updates the member list accordingly. This means that demotion will work
correctly even if there is a node failure or leader change in between
the two steps.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>