raft: Fix race when proposal is cancelled and triggered at the same time
The demotion changes revealed an interesting race. If a proposal is
triggered at the same moment that something cancels it, the callback can
be called (which commits the transaction to memory), but an error can be
returned from ProposeValue, which causes the transaction to be aborted.

ProposeValue should return an error if and only if the callback was not
called.

Add a return value to cancel(). If cancel fails because the proposal was
already triggered, avoid returning an error.

Also, remove the applyResult type, since it isn't useful.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>