raft: Don't panic on Campaign failure
Campaign used to be passed a context that was derived from Background
(n.Ctx). Awhile ago, n.Ctx was removed, and this call to Campaign was
changed to use the context passed to Run.

This context may be cancelled before Campaign is complete. This leads to
a panic.

Remove the panic, because it can trigger accidentally and it serves no
purpose.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>