swarmctl: Match full name, not name prefix
The swarmctl tool resolves names with the NamePrefixes filter. This
doesn't seem right. The user should give full names, not abbreviated
ones. It causes problems with services whose names share prefixes - for
example it's difficult to remove a service named "busybox" when one
named "busybox-top" also exists.

A case could be made for supporting ID prefixes instead, but this
isn't currently implemented, so it doesn't seem like a priority.

Use Name instead of NamePrefix to switch to exact name matching.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>