dispatcher: Correct problems with starting and stopping
The Stop method did not wait for Run to complete, so in theory it would
be possible to call Run again before the first instance is complete. Fix
this by adding a waitgroup.

"d.ctx" is used in some places without locking. In a fast restart
scenario, this could cause the wrong context to be used. Make sure that
all uses of "d.ctx" use locking.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>