ca: Remove unneeded code to cancel requests to external CAs
It shouldn't be necessary to manually close the request body when the
context is cancelled. ctxhttp should take care of this automatically.

Closing the body triggers a race on Go < 1.7, since it turns out ctxhttp
uses a wrapper around Body that is not reentrant.

Remove the goroutine that closes the body when the context is canceled.

Signed-off-by: Aaron Lehmann <aaron.lehmann@docker.com>