Update the documentation on the deprecated `ServiceSpec.Networks`
The field `ServiceSpec.Networks` has been deprecated and is replaced
by `ServiceSpec.Task.Networks`. There is no description for that
field other than a tag of `[deprecated=true]`.

This fix updates the documentation on the deprecated
`ServiceSpec.Networks` and directs user to use `ServiceSpec.Task.Networks`
instead.

This fix is related to the comment:
https://github.com/docker/docker/pull/29463#issuecomment-271387898

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>