Fix issue in service update of published ports in host mode
This fix tries to fix the issue in docker/docker/30199
where service update will not work when published ports is
host publish mode.

The reason for the issue is that `isPortsAllocated()`
does not capture host publish mode.

This fix address the issue by updating allocated ports if
there is any change in ports of host (non ingress) mode as well.

Unit tests have been added.

Signed-off-by: Yong Tang <yong.tang.github@outlook.com>